+++ data_provider/gqn_provider.py
@@ -2,6 +2,8 @@
 Input pipeline (tf.dataset and input_fn) for GQN datasets.
 Adapted from the implementation provided here:
 https://github.com/deepmind/gqn-datasets/blob/acca9db6d9aa7cfa4c41ded45ccb96fecc9b272e/data_reader.py
+
+Minimal data reader for GQN TFRecord datasets.
 """
 
 from __future__ import absolute_import
@@ -12,8 +14,7 @@ import collections
 import os
 import tensorflow as tf
 
-
-# ---------- ad-hoc data structures ----------
+from tensorflow.python.util import nest
 
 DatasetInfo = collections.namedtuple(
     "DatasetInfo",
@@ -23,8 +24,9 @@ Context = collections.namedtuple("Context", ["frames", "cameras"])
 Query = collections.namedtuple("Query", ["context", "query_camera"])
 TaskData = collections.namedtuple("TaskData", ["query", "target"])
 
-
-# ---------- dataset constants ----------
+_NUM_CHANNELS = 3
+_NUM_RAW_CAMERA_PARAMS = 5
+_MODES = ("train", "test")
 
 _DATASETS = dict(
     jaco=DatasetInfo(
@@ -51,14 +53,6 @@ _DATASETS = dict(
         frame_size=64,
         sequence_size=10,
     ),
-    # super-small subset of rooms_ring for debugging purposes
-    rooms_ring_camera_debug=DatasetInfo(
-        basepath="rooms_ring_camera_debug",
-        train_size=1,
-        test_size=1,
-        frame_size=64,
-        sequence_size=10,
-    ),
     rooms_free_camera_no_object_rotations=DatasetInfo(
         basepath="rooms_free_camera_no_object_rotations",
         train_size=2160,
@@ -81,12 +75,14 @@ _DATASETS = dict(
         sequence_size=15,
     ),
 )
-_NUM_CHANNELS = 3
-_NUM_RAW_CAMERA_PARAMS = 5
-_MODES = ("train", "test")
 
 
-# ---------- helper functions ----------
+def _get_dataset_files(dateset_info, mode, rootdir):
+    """Generates lists of files for a given dataset version."""
+    basepath = dateset_info.basepath
+    base = os.path.join(rootdir, basepath, mode)
+
+    return [os.path.join(base, f) for f in os.listdir(base)]
 
 
 def _convert_frame_data(jpeg_data):
@@ -94,117 +90,9 @@ def _convert_frame_data(jpeg_data):
     return tf.image.convert_image_dtype(decoded_frames, dtype=tf.float32)
 
 
-def _get_dataset_files(dataset_info, mode, root):
-    """Generates lists of files for a given dataset version."""
-    basepath = dataset_info.basepath
-    base = os.path.join(root, basepath, mode)
-    if mode == "train":
-        num_files = dataset_info.train_size
-    else:
-        num_files = dataset_info.test_size
-    length = len(str(num_files))
-    template = "{:0%d}-of-{:0%d}.tfrecord" % (length, length)
-    record_paths = [  # indexing runs from 1 to n
-        os.path.join(base, template.format(i, num_files))
-        for i in range(1, num_files + 1)
-    ]
-    return record_paths
+class EagerDataReader(object):
+    """Minimal eager TFRecord reader for Tensorflow >2.0b.
 
-
-def _get_randomized_indices(context_size, dataset_info):
-    """Generates randomized indices into a sequence of a specific length."""
-    example_size = context_size + 1
-    indices = tf.range(0, dataset_info.sequence_size)
-    indices = tf.random_shuffle(indices)
-    indices = tf.slice(indices, begin=[0], size=[example_size])
-    return indices
-
-
-def _parse(raw_data, dataset_info):
-    """Parses raw data from the tfrecord."""
-    feature_map = {
-        "frames": tf.FixedLenFeature(shape=dataset_info.sequence_size, dtype=tf.string),
-        "cameras": tf.FixedLenFeature(
-            shape=[dataset_info.sequence_size * _NUM_RAW_CAMERA_PARAMS],
-            dtype=tf.float32,
-        ),
-    }
-    # example = tf.parse_example(raw_data, feature_map)
-    example = tf.parse_single_example(raw_data, feature_map)
-    return example
-
-
-def _preprocess(example, indices, context_size, custom_frame_size, dataset_info):
-    """Preprocesses the parsed data."""
-    # frames
-    example_size = context_size + 1
-    frames = tf.concat(example["frames"], axis=0)
-    frames = tf.gather(frames, indices, axis=0)
-    frames = tf.map_fn(
-        _convert_frame_data, tf.reshape(frames, [-1]), dtype=tf.float32, back_prop=False
-    )
-    dataset_image_dimensions = tuple([dataset_info.frame_size] * 2 + [_NUM_CHANNELS])
-    frames = tf.reshape(frames, (example_size,) + dataset_image_dimensions)
-    if custom_frame_size and custom_frame_size != dataset_info.frame_size:
-        frames = tf.reshape(frames, dataset_image_dimensions)
-        new_frame_dimensions = (custom_frame_size,) * 2 + (_NUM_CHANNELS,)
-        frames = tf.image.resize_bilinear(
-            frames, new_frame_dimensions[:2], align_corners=True
-        )
-        frames = tf.reshape(frames, (-1, example_size) + new_frame_dimensions)
-    # cameras
-    raw_pose_params = example["cameras"]
-    raw_pose_params = tf.reshape(
-        raw_pose_params, [dataset_info.sequence_size, _NUM_RAW_CAMERA_PARAMS]
-    )
-    raw_pose_params = tf.gather(raw_pose_params, indices, axis=0)
-    pos = raw_pose_params[:, 0:3]
-    yaw = raw_pose_params[:, 3:4]
-    pitch = raw_pose_params[:, 4:5]
-    cameras = tf.concat(
-        [pos, tf.sin(yaw), tf.cos(yaw), tf.sin(pitch), tf.cos(pitch)], axis=-1
-    )
-    # return preprocessed tuple
-    preprocessed_example = {}
-    preprocessed_example["frames"] = frames
-    preprocessed_example["cameras"] = cameras
-    return preprocessed_example
-
-
-def _prepare(preprocessed_example):
-    """Prepares the preprocessed data into (feature, label) tuples."""
-    # decompose
-    frames = preprocessed_example["frames"]
-    cameras = preprocessed_example["cameras"]
-    # split data
-    context_frames = frames[:-1]
-    context_cameras = cameras[:-1]
-    target = frames[-1]
-    query_camera = cameras[-1]
-    context = Context(cameras=context_cameras, frames=context_frames)
-    query = Query(context=context, query_camera=query_camera)
-    data = TaskData(query=query, target=target)
-    return data, data.target
-
-
-# ---------- input_fn ----------
-
-
-def gqn_input_fn(
-    dataset_name,
-    root,
-    mode,
-    context_size,
-    batch_size=1,
-    num_epochs=1,
-    # optionally reshape frames
-    custom_frame_size=None,
-    # queue params
-    num_threads=4,
-    buffer_size=256,
-    seed=None,
-):
-    """
   Creates a tf.data.Dataset based op that returns data.
     Args:
       dataset_name: string, one of ['jaco', 'mazes', 'rooms_ring_camera',
@@ -238,57 +126,155 @@ def gqn_input_fn(
       ValueError: if the required version does not exist; if the required mode
          is not supported; if the requested context_size is bigger than the
          maximum supported for the given dataset version.
-  """
+    """
 
-    # map estimator mode key to dataset internal mode strings
-    if mode == tf.estimator.ModeKeys.TRAIN:
-        str_mode = "train"
-    else:
-        str_mode = "test"
-    # check validity of requested dataset and split
-    if dataset_name not in _DATASETS:
-        raise ValueError(
-            "Unrecognized dataset {} requested. Available datasets "
-            "are {}".format(dataset_name, _DATASETS.keys())
-        )
-    if str_mode not in _MODES:
-        raise ValueError(
-            "Unsupported mode {} requested. Supported modes "
-            "are {}".format(str_mode, _MODES)
-        )
-    # retrieve dataset parameters
-    dataset_info = _DATASETS[dataset_name]
-    if context_size >= dataset_info.sequence_size:
-        raise ValueError(
-            "Maximum support context size for dataset {} is {}, but "
-            "was {}.".format(dataset_name, dataset_info.sequence_size - 1, context_size)
-        )
-    # collect the paths to all tfrecord files
-    record_paths = _get_dataset_files(dataset_info, str_mode, root)
-    # create TFRecordDataset
-    dataset = tf.data.TFRecordDataset(
-        filenames=record_paths, num_parallel_reads=num_threads
-    )
-    # parse the data from tfrecords
-    dataset = dataset.map(
-        lambda raw_data: _parse(raw_data, dataset_info), num_parallel_calls=num_threads
-    )
-    # preprocess into context and target
-    indices = _get_randomized_indices(context_size, dataset_info)
-    dataset = dataset.map(
-        lambda example: _preprocess(
-            example, indices, context_size, custom_frame_size, dataset_info
-        ),
-        num_parallel_calls=num_threads,
-    )
-    # parse into tuple expected by tf.estimator input_fn
-    dataset = dataset.map(_prepare, num_parallel_calls=num_threads)
-    # shuffle data
-    if mode == tf.estimator.ModeKeys.TRAIN:
+    def __init__(
+        self,
+        dataset,
+        context_size,
+        rootdir,
+        mode="train",
+        batch_size=1,
+        num_epochs=1,
+        # Optionally reshape frames
+        custom_frame_size=None,
+        # Optionally control dataset object
+        num_threads=4,
+        buffer_size=256,
+        seed=None,
+    ):
+
+        if dataset not in _DATASETS:
+            raise ValueError(
+                "Unrecognized dataset {} requested. Available datasets "
+                "are {}".format(dataset, _DATASETS.keys())
+            )
+
+        # Dataset description
+        self._dataset_info = _DATASETS[dataset]
+
+        if mode not in _MODES:
+            raise ValueError(
+                "Unsupported mode {} requested. Supported modes are {}".format(
+                    mode, _MODES
+                )
+            )
+
+        if context_size >= self._dataset_info.sequence_size:
+            raise ValueError(
+                "Maximum support context size for dataset {} is {}, but was {}.".format(
+                    dataset, self._dataset_info.sequence_size - 1, context_size
+                )
+            )
+
+        self.seed = seed
+        self._context_size = context_size
+
+        # Number of views in the context + target view
+        self._example_size = context_size + 1
+        self._custom_frame_size = custom_frame_size
+
+        self.indices = self._get_randomized_indices()
+
+        filenames = _get_dataset_files(self._dataset_info, mode, rootdir)
+
+        # Load the dataset from files
+        dataset = tf.data.TFRecordDataset(filenames, num_parallel_reads=num_threads)
         dataset = dataset.shuffle(buffer_size=(buffer_size * batch_size), seed=seed)
-    # set up batching
-    dataset = dataset.repeat(num_epochs)
-    dataset = dataset.batch(batch_size)
-    dataset = dataset.prefetch(buffer_size)
-    return dataset
+        dataset = dataset.map(self._parse_record, num_parallel_calls=num_threads)
+        dataset = dataset.map(self._preprocess, num_parallel_calls=num_threads)
+        dataset = dataset.map(self._prepare, num_parallel_calls=num_threads)
+        dataset = dataset.repeat(num_epochs)
+        dataset = dataset.batch(batch_size)
+        dataset = dataset.prefetch(buffer_size)
 
+        self.dataset = dataset
+
+    def _prepare(self, frames, cameras):
+        """Reads batch_size (query, target) pairs."""
+        context_frames = frames[:-1]
+        context_cameras = cameras[:-1]
+        target = frames[-1]
+        query_camera = cameras[-1]
+        context = Context(cameras=context_cameras, frames=context_frames)
+        query = Query(context=context, query_camera=query_camera)
+
+        return TaskData(query=query, target=target)
+
+    def _parse_record(self, record):
+        feature_map = {
+            "frames": tf.io.FixedLenFeature(
+                shape=self._dataset_info.sequence_size, dtype=tf.string
+            ),
+            "cameras": tf.io.FixedLenFeature(
+                shape=[self._dataset_info.sequence_size * _NUM_RAW_CAMERA_PARAMS],
+                dtype=tf.float32,
+            ),
+        }
+        example = tf.io.parse_single_example(record, feature_map)
+
+        return example
+
+    def _preprocess(self, example):
+        """Preprocess raw frames and cameras from parsed tfrecord"""
+
+        frames = self._preprocess_frames(
+            example,
+            self.indices,
+            self._example_size,
+            self._dataset_info,
+            self._custom_frame_size,
+        )
+        cameras = self._preprocess_cameras(example, self.indices, self._dataset_info)
+
+        return frames, cameras
+
+    def _get_randomized_indices(self):
+        """Generates randomized indices into a sequence of a specific length."""
+
+        indices = tf.range(0, self._dataset_info.sequence_size)
+        indices = tf.random.shuffle(indices, seed=self.seed)
+        indices = tf.slice(indices, begin=[0], size=[self._example_size])
+        return indices
+
+    def _preprocess_frames(
+        self, example, indices, example_size, dataset_info, custom_frame_size
+    ):
+        """Instantiates the ops used to preprocess the frames data."""
+
+        frames = tf.concat(example["frames"], axis=0)
+        frames = tf.gather(frames, indices, axis=0)
+        frames = tf.map_fn(
+            _convert_frame_data,
+            tf.reshape(frames, [-1]),
+            dtype=tf.float32,
+            back_prop=False,
+        )
+        dataset_image_dimensions = tuple(
+            [dataset_info.frame_size] * 2 + [_NUM_CHANNELS]
+        )
+        frames = tf.reshape(frames, (example_size,) + dataset_image_dimensions)
+        if custom_frame_size and custom_frame_size != dataset_info.frame_size:
+            frames = tf.reshape(frames, dataset_image_dimensions)
+            new_frame_dimensions = (custom_frame_size,) * 2 + (_NUM_CHANNELS,)
+            frames = tf.image.resize_bilinear(
+                frames, new_frame_dimensions[:2], align_corners=True
+            )
+            frames = tf.reshape(frames, (-1, example_size) + new_frame_dimensions)
+
+        return frames
+
+    def _preprocess_cameras(self, example, indices, dataset_info):
+        """Instantiates the ops used to preprocess the cameras data."""
+        raw_pose_params = example["cameras"]
+        raw_pose_params = tf.reshape(
+            raw_pose_params, [dataset_info.sequence_size, _NUM_RAW_CAMERA_PARAMS]
+        )
+        raw_pose_params = tf.gather(raw_pose_params, indices, axis=0)
+        pos = raw_pose_params[:, 0:3]
+        yaw = raw_pose_params[:, 3:4]
+        pitch = raw_pose_params[:, 4:5]
+        cameras = tf.concat(
+            [pos, tf.sin(yaw), tf.cos(yaw), tf.sin(pitch), tf.cos(pitch)], axis=-1
+        )
+        return cameras
